refined dTheta correction with cosine

+++ PoseTransfer.java
@@ -7,7 +7,7 @@ public class PoseTransfer {
     public static Pose2d currentPose = new Pose2d(-59.9,-59.9,0);
     public static int idealGridCoordX = 0;
     public static int idealGridCoordY = 0;
-    public static int idealGridAngle = 0;
+    public static double idealGridAngle = 0;
     public static double slidePos = 0;
 
     public static boolean isBlue = true;
